Move recommendations response type
