+++ packages/headless/src/api/commerce/recommendations/recommendations-response.ts
@@ -1,0 +1,6 @@
+import {BaseCommerceSuccessResponse} from '../common/response';
+
+export interface RecommendationsCommerceSuccessResponse
+  extends BaseCommerceSuccessResponse {
+  headline: string;
+}
